+++ src/README.md
@@ -1,17 +1,12 @@
 
 ## Directories
 
+Source files for efu, detector modules and data generators.
+
 Directory             | Function
 -------------         | -------------
-adc_readout           | detector implementation
-gdgem                 | detector implementation
-multiblade            | detector implementation
-multigrid             | detector implementation
-sonde                 | detector implementation
 common                | common detector code
-readout               | common readout data parser
 efu                   | EFU main application code
-generators            | data generators - replays previously recorded data
-test                  | unit test utilities
-tools                 | replay of network (pcap) traffic
-udp                   | UDP receiver 'detector' - obsolete
+modules               | detector implementations
+test                  | unit test utilities:
+udpgenpcap            | replay of network (pcap) traffic
