+++ src/modules/multigrid/mesytec/test/CMakeLists.txt
@@ -51,11 +51,11 @@ endif()
 set(TEST_DATA_PATH2 "${REFDATA}/multigrid/2018_08_30")
 if(EXISTS ${TEST_DATA_PATH2})
   set(MGBuilderMesytecTest_SRC
-    ${ESS_SOURCE_DIR}/tools/ReaderPcap.cpp
+    ${ESS_SOURCE_DIR}/udpgenbpcap/ReaderPcap.cpp
     BuilderMesytecTest.cpp
     )
   set(MGBuilderMesytecTest_INC
-    ${ESS_SOURCE_DIR}/tools/ReaderPcap.h
+    ${ESS_SOURCE_DIR}/udpgenpcap/ReaderPcap.h
     TestData.h
     )
   set(MGBuilderMesytecTest_LIB
