--- utils/centos/tune/setvalues_default
@@ -1,31 +1,0 @@
-#!/bin/bash
-
-function syssettings()
-{
-  sysctl -a | grep 'rmem_max\|wmem_max\|netdev_max_backlog\|netdev_budget'
-  ip link show | grep "eno49\|eno50"
-  ethtool -g eno49
-  ethtool -c eno49
-  ethtool -g eno50
-  ethtool -c eno50
-}
-
-echo System settings before change
-syssettings
-
-echo changing values
-sysctl -w net.core.rmem_max=212992
-sysctl -w net.core.wmem_max=212992
-sysctl -w net.core.netdev_max_backlog=1000
-sysctl -w net.core.netdev_budget=300
-
-ifconfig eno49 mtu 9000 txqueuelen 1000 up
-ifconfig eno50 mtu 9000 txqueuelen 1000 up
-
-ethtool -G eno49 rx 453 tx 4078
-ethtool -G eno50 rx 453 tx 4078
-ethtool -C eno49 rx-usecs 24 
-ethtool -C eno50 rx-usecs 24 
-
-echo System settings after change
-syssettings
