--- utils/centos/tune/setvalues_optimized2
@@ -1,31 +1,0 @@
-#!/bin/bash
-
-function syssettings()
-{
-  sysctl -a | grep 'rmem_max\|wmem_max\|netdev_max_backlog\|netdev_budget'
-  ip link show | grep "eno49\|eno50"
-  ethtool -g eno49
-  ethtool -c eno49
-  ethtool -g eno50
-  ethtool -c eno50
-}
-
-echo System settings before change
-syssettings
-
-echo changing values
-sysctl -w net.core.rmem_max=12582912
-sysctl -w net.core.wmem_max=12582912
-sysctl -w net.core.netdev_max_backlog=250000
-sysctl -w net.core.netdev_budget=600
-
-ifconfig eno49 mtu 9000 txqueuelen 10000 up
-ifconfig eno50 mtu 9000 txqueuelen 10000 up
-
-ethtool -G eno49 rx 4078 tx 4078 
-ethtool -G eno50 rx 4078 tx 4078
-ethtool -C eno49 rx-usecs 48 
-ethtool -C eno50 rx-usecs 48 
-
-echo System settings after change
-syssettings
